+++ src/pl-utf8.h
@@ -55,6 +55,6 @@
 extern char *_PL__utf8_get_char(const char *in, int *chr);
 extern char *_PL__utf8_put_char(char *out, int chr);
 
-extern unsigned int utf8_strlen(const char *s, unsigned int len);
+extern size_t utf8_strlen(const char *s, size_t len);
 
 #endif /*UTF8_H_INCLUDED*/
